[REFACTOR] Service 작업을 Impl로 분리
